server/requirements.txt
-i https://pypi.org/simple
alembic==1.16.4; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
flask==3.1.1; python_version >= '3.9'
flask-cors==6.0.1; python_version >= '3.9' and python_version < '4.0'
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
flask-migrate==4.1.0; python_version >= '3.6'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.2.4; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
passlib==1.7.4
psycopg2-binary==2.9.9; python_version >= '3.7'
pyjwt==2.10.1; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
pytz==2024.2
setuptools==70.3.0; python_version >= '3.8'
sqlalchemy==2.0.29; python_version >= '3.7'
sqlalchemy-serializer==1.4.22; python_version >= '3.10' and python_version < '4.0'
typing-extensions==4.14.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
gunicorn

